+++ ee/hogai/tools/__init__.py
@@ -3,13 +3,13 @@ from .create_form import CreateFormTool
 from .create_insight import CreateInsightTool
 
 # MCP tool imports trigger @mcp_tool_registry.register decorators
-from .execute_sql.external import ExecuteSQLMCPTool  # noqa: F401
+from .execute_sql.mcp_tool import ExecuteSQLMCPTool  # noqa: F401
 from .execute_sql.tool import ExecuteSQLTool
 from .list_data import ListDataTool
 from .manage_memories import ManageMemoriesTool
 from .read_data import ReadDataTool
-from .read_data_warehouse_schema_external import ReadDataWarehouseSchemaMCPTool  # noqa: F401
-from .read_taxonomy.external import ReadTaxonomyMCPTool  # noqa: F401
+from .read_data_warehouse_schema_mcp_tool import ReadDataWarehouseSchemaMCPTool  # noqa: F401
+from .read_taxonomy.mcp_tool import ReadTaxonomyMCPTool  # noqa: F401
 from .read_taxonomy.tool import ReadTaxonomyTool
 from .search import SearchTool
 from .switch_mode import SwitchModeTool
