+++ ee/hogai/tools/execute_sql/mcp_tool.py
@@ -5,16 +5,16 @@ from pydantic import BaseModel, Field
 from ee.hogai.chat_agent.schema_generator.parsers import PydanticOutputParserException
 from ee.hogai.chat_agent.sql.mixins import HogQLOutputParserMixin
 from ee.hogai.context.insight.context import InsightContext
-from ee.hogai.external_tool import MCPTool, mcp_tool_registry
+from ee.hogai.mcp_tool import MCPTool, mcp_tool_registry
 from ee.hogai.tool_errors import MaxToolRetryableError
 
 
-class ExecuteSQLExternalToolArgs(BaseModel):
+class ExecuteSQLMCPToolArgs(BaseModel):
     query: str = Field(description="The final SQL query to be executed.")
 
 
 @mcp_tool_registry.register(scopes=["insight:read", "query:read"])
-class ExecuteSQLMCPTool(HogQLOutputParserMixin, MCPTool[ExecuteSQLExternalToolArgs]):
+class ExecuteSQLMCPTool(HogQLOutputParserMixin, MCPTool[ExecuteSQLMCPToolArgs]):
     """
     MCP version of ExecuteSQLTool.
 
@@ -22,9 +22,9 @@ class ExecuteSQLMCPTool(HogQLOutputParserMixin, MCPTool[ExecuteSQLExternalToolAr
     """
 
     name = "execute_sql"
-    args_schema = ExecuteSQLExternalToolArgs
+    args_schema = ExecuteSQLMCPToolArgs
 
-    async def execute(self, args: ExecuteSQLExternalToolArgs) -> tuple[str, dict[str, Any] | None]:
+    async def execute(self, args: ExecuteSQLMCPToolArgs) -> tuple[str, dict[str, Any] | None]:
         try:
             validated_query = await self._validate_hogql_query(args.query)
         except PydanticOutputParserException as e:
